Added a few message handlers to the server logic

+++ messages.h
@@ -59,17 +59,17 @@ uint8_t get_header_type(char *msg);
 const char *get_header_type_string(uint8_t resource);
 
  /* REQUESTS */
-void create_msg_get_block_request(uuid_t uuid, char **msg, uint32_t *msg_len);
-void create_msg_get_file_request(char *filename, char **msg, uint32_t *msg_len);
+void create_msg_get_block_request(uuid_t uuid, char **msg, uint64_t *msg_len);
+void create_msg_get_file_request(char *filename, char **msg, uint64_t *msg_len);
 void create_msg_get_folder_meta_request(char *dirname, char **msg, 
-                                        uint32_t *msg_len);
-void create_msg_get_servers_request(char **msg, uint32_t *msg_len);
+                                        uint64_t *msg_len);
+void create_msg_get_servers_request(char **msg, uint64_t *msg_len);
 void create_msg_post_block_request(uuid_t uuid, uint32_t size, char *content,
-                                   char **msg, uint32_t *msg_len);
+                                   char **msg, uint64_t *msg_len);
 void create_msg_post_file_request(char *filename, uint64_t file_size, 
-                                  char **msg, uint32_t *msg_len);
+                                  char **msg, uint64_t *msg_len);
 void create_msg_post_folder_request(char *foldername, char **msg, 
-                                    uint32_t *msg_len);
+                                    uint64_t *msg_len);
 void parse_msg_get_block_request(char *msg, uuid_t *uuid);
 void parse_msg_get_file_request(char *msg, char **filename);
 void parse_msg_get_folder_meta_request(char *msg, char **dirname);
@@ -80,15 +80,15 @@ void parse_msg_post_folder_request(char *msg, char **foldername);
 
 /* RESPONSES */
 void create_msg_get_block_response(uuid_t uuid, uint32_t size, char *content,
-                                   char **msg, uint32_t *msg_len);
+                                   char **msg, uint64_t *msg_len);
 void create_msg_get_file_response(meta_t *file_meta, char **msg, 
-                                  uint32_t *msg_len);
+                                  uint64_t *msg_len);
 void create_msg_get_folder_meta_response(meta_t *folder, char **msg, 
-                                         uint32_t *msg_len);
-void create_msg_get_servers_response(char **msg, uint32_t *msg_len);
-void create_msg_post_block_response(uuid_t uuid, char **msg, uint32_t *msg_len);
-void create_msg_post_file_response(meta_t *file, char **msg, uint32_t *msg_len);
-void create_msg_post_folder_response(char **msg, uint32_t *msg_len);
+                                         uint64_t *msg_len);
+void create_msg_get_servers_response(char **msg, uint64_t *msg_len);
+void create_msg_post_block_response(uuid_t uuid, char **msg, uint64_t *msg_len);
+void create_msg_post_file_response(meta_t *file, char **msg, uint64_t *msg_len);
+void create_msg_post_folder_response(char **msg, uint64_t *msg_len);
 void parse_msg_get_block_response(char *msg, uuid_t *uuid, uint32_t *size, 
                                   char **content);
 void parse_msg_get_file_response(char *msg, basic_block_t **blocks, 
@@ -106,10 +106,10 @@ void parse_msg_post_folder_response(char *msg);
  * ================ */ 
 void msg_with_block(uint8_t action, uint8_t resource, uint8_t type,
                     uuid_t uuid, uint32_t size, char *content,
-                    char **msg, uint32_t *msg_len);
+                    char **msg, uint64_t *msg_len);
 void msg_with_single_element(uint8_t action, uint8_t resource, uint8_t type, 
                              void *element, uint32_t element_len, char **msg, 
-                             uint32_t *msg_len);
+                             uint64_t *msg_len);
 void parse_msg_block(char *msg, uint8_t action, uint8_t resource, uint8_t type, 
                      uuid_t *uuid, uint32_t *size, char **content);
 void parse_msg_single_string(char *msg, uint8_t action, uint8_t resource,
